* Example platforms:    * Rework example software  * Headers    * IO helpers  * SystemC    * Conform to new offset definitions, have a correct exception display  * Soclib-CC    * Let the user choose the output filename
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@4 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ systemc/src/common/address_decoding_table.cc
@@ -112,7 +112,7 @@ tmpl(void)::print( std::ostream &o ) const
 
 tmpl(bool)::isAllBelow( output_t val ) const
 {
-    for ( size_t i=0; i<(1<<m_use_bits); ++i )
+    for ( size_t i=0; i<(size_t)(1<<m_use_bits); ++i )
         if ( m_table[i] >= val )
             return false;
     return true;
